Split libsql workspace module
Decompose the oversized libSQL workspace module into focused
submodules for vector search, FTS, document operations, chunk
operations, and tests.

Keep the NativeWorkspaceStore implementation in workspace/mod.rs
while delegating the heavier CRUD logic to internal helpers so every
workspace module file stays under the repository's 400-line ceiling
without changing behaviour or signatures.
